Remove change_detector.rb
Fixes #2.

Apparently this was a work-in-progress in 2011. However, I don't
appear to have done any further work on it since then, so I've created
a branch [1] to capture the state of the code at this point and then
deleted the file in this commit. I don't believe it is used anywhere
and all the tests still pass, so it should be safe to delete.

[1]: https://github.com/floehopper/introspection/tree/change-detector